agentic_rag_test/requirements.txt
# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 环境变量
python-dotenv>=1.0.0

# 大模型与 Embedding（OpenAI 兼容 API）
openai>=1.12.0
langchain-deepseek>=0.0.6
deepagents>=0.1.0

# 向量数据库
qdrant-client>=1.7.0

# 文档解析
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-docx>=1.1.0
python-pptx>=0.6.21
docx2pdf>=0.1.8

# 数据库
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
